Correcting location of test efficiency and merging with the previous version of the test file.
